requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-import-export==2.3.0
et-xmlfile==1.0.1
jdcal==1.4.1
joblib==0.16.0
kiwisolver==1.2.0
MarkupPy==1.14
matplotlib==3.3.1
numpy==1.19.1
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.10.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
tablib==2.0.0
threadpoolctl==2.1.0
xlrd==1.2.0
xlwt==1.3.0
